requirements.txt
streamlit  == 1.24.0
datetime
Pillow ==9.5.0
streamlit-option-menu ==0.3.6